@@ -17,48 +17,53 @@
 
 source "https://rubygems.org"
 
-gem "rails", "2.3.17"
+gem "rails", "~> 4.1"
+gem "haml-rails", "~> 0.5"
+gem "sass-rails", "~> 4.0"
+gem "rainbows-rails", "~> 1.0"
+gem "ember-rails", "~> 0.15"
 
-gem "rack", "1.1.6"
-gem "app_config", "1.0.2"
-gem "haml", "3.1.6"
-gem "hike", "1.2.1"
-gem "i18n", "0.4.2"
-gem "json", "1.6.1"
-gem "kwalify", "0.7.2"
-gem "multi_json", "1.0.3"
-gem "sass", "3.2.12"
-gem "simple-navigation", "3.7.0"
-gem "sprockets", "2.10.1"
-gem "sprockets-sass", "1.0.2"
-gem "sprockets-helpers", "1.1.0"
-gem "sqlite3", "1.3.6"
-gem "syslogger", "1.3.0"
-gem "tilt", "1.3.3"
-gem "mime-types", "1.18", :require => "mime/types"
+gem "activeresource", "~> 4.0"
+gem "dotenv", "~> 1.0"
+gem "hashie", "~> 3.3"
+gem "i18n-js", "~> 2.1"
+gem "js-routes", "~> 0.9"
+gem "kwalify", "~> 0.7"
+gem "redcarpet", "~> 3.2"
+gem "redis", "~> 3.1"
+gem "sidekiq", "~> 3.2"
+gem "simple-navigation", "~> 3.12"
+gem "simple_navigation_renderers", "~> 1.0"
+gem "sqlite3", "~> 1.3"
+gem "syslogger", "~> 1.6"
+gem "uglifier", "~> 2.2"
 
-gem "chef", "~> 10.24.0"
-gem "mixlib-shellout", "~> 1.1.0"
-gem "ohai", "~> 6.14.0"
+gem "ohai", "~> 6.22"
+gem "chef", "~> 10.24"
 
-group :development do
-  gem "brakeman", "1.8.2"
-  gem "rspec-rails", "1.3.4"
+gem "mixlib-shellout", "~> 1.4",
+  require: "mixlib/shellout"
+
+gem "activerecord-session_store", "~> 0.1",
+  require: "activerecord/session_store"
+
+gem "mime-types", "~> 1.25",
+  require: "mime/types"
+
+gem "dotenv-deployment", "~> 0.2",
+  require: false
+
+gem "rack-mini-profiler", "~> 0.9",
+  require: false
+
+group :development, :test do
+  gem "brakeman", "~> 2.6"
+  gem "rspec-rails", "~> 2.14"
 end
 
 group :test do
-  gem "rspec-rails", "1.3.4"
-  gem "mocha", "0.10.5"
-  gem "sinatra", "1.2.9"
-  gem "webmock", "1.16.1"
-  gem "rcov", "1.0.0"
-end
-
-# Install gems from each barclamp. The crowbar_path variable is
-# needed by Gemfile.d/*.gemfile so that they can determine the correct
-# path at run-time rather than having to hard-code a path at
-# install-time which may be different and therefore incorrect.
-crowbar_path = File.expand_path("..", __FILE__)
-Dir.glob(File.join(crowbar_path, "Gemfile.d", "*.gemfile")) do |gemfile|
-  eval(IO.read(gemfile), binding)
+  gem "mocha", "~> 1.1"
+  gem "sinatra", "~> 1.4"
+  gem "webmock", "~> 1.19"
+  gem "simplecov", "~> 0.9"
 end
