@@ -15,8 +15,6 @@
 # limitations under the License.
 #
 
-ActionController::Base.session = {
-  :key => "_crowbar_framework_session",
-  :secret => "ae66894fcee606d20eab9637a28684a8e9a12d36a91d075035e5703ed49bb26a0f9163fd0954185dea2b701cf79cefad152fb6da0075af43066b790707f52424"
-}
-ActionController::Base.session_store = :active_record_store
+Rails.application.config.tap do |config|
+  config.session_store :active_record_store, key: "crowbar"
+end
