@@ -28,7 +28,8 @@ class NodeObject < ChefObject
     else
       ChefObject.query_chef.search "node", "#{chef_escape(search)}"
     end
-    if nodes[2] != 0 and !nodes[0].nil?
+
+    if nodes.is_a?(Array) and nodes[2] != 0 and !nodes[0].nil?
       nodes[0].delete_if { |x| x.nil? }
       answer = nodes[0].map do |x|
         NodeObject.new x
@@ -343,7 +344,7 @@ class NodeObject < ChefObject
     @role.description = chef_description
 
     # move this to event driven model one day
-    system("sudo", "-i", Rails.root.join("..", "bin", "single_chef_client.sh").expand_path)
+    system("sudo", "-i", Rails.root.join("..", "bin", "single_chef_client.sh").expand_path.to_s)
   end
 
   def public_name(suggest=false)
