@@ -44,6 +44,10 @@ module Dsl
       end
 
       def header(dep_raw = true, attr_raw = false)
+        raw_options = {}
+        raw_options[:dep_raw] = true if dep_raw
+        raw_options[:attr_raw] = true if attr_raw
+
         content_tag(
           :h3,
           [
@@ -52,11 +56,11 @@ module Dsl
               :div,
               content_tag(
                 :small,
-                proposal_raw_button(proposal, :dep_raw => dep_raw, :attr_raw => attr_raw)
+                proposal_raw_button(proposal, raw_options)
               ),
               :class => "pull-right"
             )
-          ].join("\n")
+          ].join("\n").html_safe
         )
       end
 
